Improved naming of rollup plugin to pick values from package.json
