+++ .eslintrc.yml
@@ -48,7 +48,7 @@ overrides:
   - files:
       - esm-utils.js
       - rollup.config.js
-      - scripts/rollup-manifest-filter.js
+      - scripts/pick-from-package-json.js
     parserOptions:
       ecmaVersion: 2018
       sourceType: module
